third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/popup-same-origin-allow-popups-with-same-site.https-expected.txt

This is a testharness.js-based test.
PASS same-origin-allow-popups document opening popup to https://www1.web-platform.test:8444 with COOP: ""
PASS same-origin-allow-popups document opening popup to https://www1.web-platform.test:8444 with COOP: "unsafe-none"
PASS same-origin-allow-popups document opening popup to https://www1.web-platform.test:8444 with COOP: "jibberish"
PASS same-origin-allow-popups document opening popup to https://www1.web-platform.test:8444 with COOP: "same-site"
PASS same-origin-allow-popups document opening popup to https://www1.web-platform.test:8444 with COOP: "same-site unsafe-allow-outgoing"
PASS same-origin-allow-popups document opening popup to https://www1.web-platform.test:8444 with COOP: "same-origin unsafe-allow-outgoing"
FAIL same-origin-allow-popups document opening popup to https://www1.web-platform.test:8444 with COOP: "same-origin" assert_equals: name expected "" but got "same-origin-allow-popups_to_SAME_SITE_same-origin"
FAIL same-origin-allow-popups document opening popup to https://www1.web-platform.test:8444 with COOP: "same-origin-allow-popups" assert_equals: name expected "" but got "same-origin-allow-popups_to_SAME_SITE_same-origin-allow-popups"
Harness: the test ran to completion.
